chore: uint256 casting on creditCerUsd

+++ CerbySwapV1_Tests/contracts/CerbySwapV1_SwapFunctions.sol
@@ -415,7 +415,7 @@ abstract contract CerbySwapV1_SwapFunctions is CerbySwapV1_LiquidityFunctions {
         Pool storage pool = pools[tokenToPoolId[_token]];
         if (
             pool.creditCerUsd < MAX_CER_USD_CREDIT &&
-            pool.creditCerUsd + amountCerUsdIn < _amountCerUsdOut
+            uint256(pool.creditCerUsd) + amountCerUsdIn < _amountCerUsdOut
         ) {
             revert("Z");
             revert CerbySwapV1_CreditCerUsdMustNotBeBelowZero();
